ai/requirements.txt
Flask==2.2.5
openai
transformers
torch
youtube_transcript_api
python-dotenv
nltk
requests
yt-dlp
